+++ packages/react-querybuilder/src/hooks/index.ts
@@ -1,6 +1,5 @@
-// Don't export these here as they need to stay internal
-// export * from './useQueryBuilder';
-// export * from './useQueryBuilderSetup';
+export * from './useQueryBuilderSchema';
+export * from './useQueryBuilderSetup';
 export * from './useRule';
 export * from './useRuleGroup';
 export * from './useSelectElementChangeHandler';