+++ packages/react-querybuilder/src/index.ts
@@ -4,6 +4,7 @@ import * as messages from './messages';
 export * from './components';
 export * from './defaults';
 export * from './hooks';
+export * from './redux';
 export * from './types';
 export * from './utils';
 export { messages };